Save the plan folder path in the settings.
Instead of using the image directory as the default plan directory,
use a separate persistent setting for the plan directory.  Also, use
this same directory for the screenshots and the reports.

For the CT, instead of starting in the same directory as the primary
dicom files, start one directory up.