Use Affine registration for secondary.
Previously, this code was written for the CerebraWML application, which used
either (1) rigid registration for subject-to-self or (2) scaling along the
target axes for subject-to-atlas, where the latter is not applicable to
Tactics.

Affine registration is used here because the secondary might be an EPI image,
which is mildly distorted and can benefit from affine registration.