Use Rigid registration (as was used before).
The recent change to affine registration was a bad choice on two counts.
First, it increased the registration time by a factor of up to 10x.
Second, when registering e.g. T1 to DTI (or to any modality where only
everything outside the cranium is dark) the registration would fail.
For patient-to-self registration, a rigid transform is much more robust.