Use a json file to save and load plan
Now, when a plan is saved, the images are saved in the same directory
as the plan file (as <prefix>_primary.nii.gz and <prefix>_secondary.nii.gz
where <prefix> is the name of the plan file sans extension).  Likewise,
when a plan is loaded, the directory is searched for these files and they
are loaded if present.

Support for loading the old plan files has been removed.